Shuffle around the plot methods. qtplot is dead.
