+++ register/visualize/dialog.py
@@ -1,13 +1,14 @@
-# -*- coding: utf-8 -*-
+""" A custom dialog for introspection of search paths."""
 
-# Form implementation generated from reading ui file 'dialog.ui'
-#
-# Created: Tue Sep 27 14:12:23 2011
-#      by: PyQt4 UI code generator 4.8.5
-#
-# WARNING! All changes made in this file will be lost!
+import numpy as np
+from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as FigureCanvas
+from matplotlib.figure import Figure
 
 from PyQt4 import QtCore, QtGui
+
+#===============================================================================
+# Custom dialog generated by qt-designer.
+#===============================================================================
 
 try:
     _fromUtf8 = QtCore.QString.fromUtf8
@@ -30,14 +31,11 @@ class Ui_Dialog(object):
         self.splitter.setMinimumSize(QtCore.QSize(400, 0))
         self.splitter.setOrientation(QtCore.Qt.Horizontal)
         self.splitter.setObjectName(_fromUtf8("splitter"))
-        
         self.widget = MyDynamicMplCanvas(self.splitter)
-      
         sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Maximum, QtGui.QSizePolicy.Preferred)
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(0)
         sizePolicy.setHeightForWidth(self.widget.sizePolicy().hasHeightForWidth())
-        
         self.widget.setSizePolicy(sizePolicy)
         self.widget.setMinimumSize(QtCore.QSize(1000, 600))
         self.widget.setObjectName(_fromUtf8("widget"))
@@ -58,25 +56,6 @@ class Ui_Dialog(object):
         self.verticalLayout_2.addWidget(self.listWidget)
         self.horizontalLayout.addWidget(self.splitter)
         
-#        
-#        self.listWidget.connect(
-#            self.listWidget, 
-#            QtCore.SIGNAL("itemDoubleClicked(QListWidgetItem*)"), 
-#            self.mouseSlot
-#            )
-#        
-#        self.listWidget.connect(
-#            self.listWidget, 
-#            QtCore.SIGNAL("itemClicked(QListWidgetItem*)"), 
-#            self.mouseSlot
-#            )
-#        
-#        self.listWidget.connect(
-#            self.listWidget, 
-#            QtCore.SIGNAL("itemActivated(QListWidgetItem*)"), 
-#            self.mouseSlot
-#            )
-#
         self.listWidget.connect(
             self.listWidget, 
             QtCore.SIGNAL("currentItemChanged (QListWidgetItem *,QListWidgetItem *)"), 
@@ -119,11 +98,6 @@ class Ui_Dialog(object):
 ###############################################################################
 # MPL widget 
 ###############################################################################
-import plot
-import numpy as np
-
-from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.figure import Figure
 
 class MyMplCanvas(FigureCanvas):
     """Ultimately, this is a QWidget (as well as a FigureCanvasAgg, etc.)."""
@@ -137,14 +111,14 @@ class MyMplCanvas(FigureCanvas):
         if step is not None:
             self.formPlot(step)
         
-        
-            
         FigureCanvas.__init__(self, self.fig)
         self.setParent(parent)
 
-        FigureCanvas.setSizePolicy(self,
-                                   QtGui.QSizePolicy.Expanding,
-                                   QtGui.QSizePolicy.Expanding)
+        FigureCanvas.setSizePolicy(
+            self,
+            QtGui.QSizePolicy.Expanding,
+            QtGui.QSizePolicy.Expanding)
+        
         FigureCanvas.updateGeometry(self)
 
 class MyDynamicMplCanvas(MyMplCanvas):
@@ -153,32 +127,31 @@ class MyDynamicMplCanvas(MyMplCanvas):
         MyMplCanvas.__init__(self, *args, **kwargs)
 
     def formPlot(self, step):
+        """
+        A graphical representation of an optimization step. 
+    
+        Parameters
+        ----------
+        step: Register.optstep
+            A step in the registration.
+        """
         
         self.img1 = self.ax1.imshow(
             step.warpedImage, 
-            #origin='lower', 
             interpolation='nearest',
             cmap='gray'
             )
         
         self.img2 = self.ax2.imshow(
             step.warp[0] - step.grid[0], 
-            #origin='lower', 
             interpolation='nearest',
             cmap='jet'
             )
       
         self.img3 = self.ax3.imshow(
             step.warp[1] - step.grid[1], 
-            #origin='lower', 
             interpolation='nearest',
             cmap='jet'
             )
        
-        self.draw()
-    
-    def updatePlot(self, step):
-        self.img1.set_data(step.warpedImage)
-        self.img2.set_data(step.warped[0])
-        self.img3.set_data(step.warped[1])
         self.draw()
